check_ticket/requirements.txt
pdfminer3k
func_timeout==4.3.0
BS4
lxml
requests